Revert "uci.conffile: associate private types with /etc/config/{fstab,wireless}"
Looking at the code i doubt that this will work consistently as uci creates backup files with random suffix (not sure why a random suffix would be needed here) and then renames them
But aside from the above this solution wasnt as good as it seemed anyway because usually entities use the uci command to parse a config file and uci needs access to them all (uci gets run with a domain transition usually). Its up to uci to decide.

This reverts commit 12e95b037325955848de8ea572bc9e46f401e90a.

Signed-off-by: Dominick Grift <dominick.grift@defensec.nl>